+++ src/impl/Utilities.cpp
@@ -9,5 +9,5 @@ double GetRunningTime()
 {
 	auto chronoTime = std::chrono::high_resolution_clock::now();
 
-	return std::chrono::duration_cast<std::chrono::nanoseconds>(chronoTime - s_ChronoStartTime).count() / 1e+9;
+	return (double)std::chrono::duration_cast<std::chrono::nanoseconds>(chronoTime - s_ChronoStartTime).count() / 1e+9;
 }
